Updated the user service and added API.
Added API routes for user.
Modified services to better match the API (and what will be used by the front).
The user model will be changed to include playlistIds owned by the user.
I also while writing and testing routes the brilliant thought of not returning the hashed password and email.
It's obvious, but when you've never wrote clear, concise and "secure" entire stack for a webapp, you never thought of that beforehand.