Final commit for resubmission
Modified the routes to now work only with the v1 folder for every requests and
root file for navigation.
Moved the authentication part to the routes instead of the controller.
Added the option to delete and modify playlists.
Added the private feature, to unlist your playlist from public searches.
Added icons to match the features listed above.